api/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Django==4.0.6
django-environ==0.4.5
django-admin-interface==0.22.1
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-gis==1.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.4
Flask==2.1.3
geos==0.2.3
idna==3.3
inflection==0.5.1
install==1.3.5
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.2.0
psycopg2-binary==2.8.6
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
pyshp==2.3.1
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.11
Werkzeug==2.1.2
xlrd==2.0.1
